+++ src/components/button/button.styles.ts
@@ -408,13 +408,7 @@ export default css`
   }
 
   .button--caret .button__caret {
-    display: flex;
-    align-items: center;
-  }
-
-  .button--caret .button__caret svg {
-    width: 1em;
-    height: 1em;
+    height: auto;
   }
 
   /*