+++ src/components/image-comparer/image-comparer.ts
@@ -30,6 +30,7 @@ import type { CSSResultGroup } from 'lit';
  * @csspart after - The container that holds the "after" image.
  * @csspart divider - The divider that separates the images.
  * @csspart handle - The handle that the user drags to expose the after image.
+ * @csspart handle-icon - The drag icon that appears inside the handle.
  *
  * @cssproperty --divider-width - The width of the dividing line.
  * @cssproperty --handle-size - The size of the compare handle.
@@ -143,13 +144,7 @@ export default class SlImageComparer extends ShoelaceElement {
             tabindex="0"
           >
             <slot name="handle-icon">
-              <svg width="24" height="24" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                <g fill="currentColor" fill-rule="nonzero">
-                  <path
-                    d="m21.14 12.55-5.482 4.796c-.646.566-1.658.106-1.658-.753V7a1 1 0 0 1 1.659-.753l5.48 4.796a1 1 0 0 1 0 1.506h.001ZM2.341 12.55l5.482 4.796c.646.566 1.658.106 1.658-.753V7a1 1 0 0 0-1.659-.753l-5.48 4.796a1 1 0 0 0 0 1.506h-.001Z"
-                  />
-                </g>
-              </svg>
+              <sl-icon part="handle-icon" library="system" name="grip-vertical"></sl-icon>
             </slot>
           </div>
         </div>